refactor: 💡 separate AppSettings enums to new file
