+++ app/src/main/java/com/rickyhu/hushkeyboard/settings/AppSettings.kt
@@ -2,6 +2,8 @@ package com.rickyhu.hushkeyboard.settings
 
 import android.content.Context
 import androidx.datastore.dataStore
+import com.rickyhu.hushkeyboard.settings.options.ThemeOption
+import com.rickyhu.hushkeyboard.settings.options.WideNotationOption
 import kotlinx.serialization.Serializable
 
 val Context.dataStore by dataStore("app-settings.json", AppSettingsSerializer)
@@ -13,22 +15,3 @@ data class AppSettings(
     val vibrateOnTap: Boolean = true,
     val wideNotationOption: WideNotationOption = WideNotationOption.WideWithW
 )
-
-enum class ThemeOption {
-    System, Light, Dark;
-
-    fun isDarkTheme(isSystemInDarkMode: Boolean) = when (this) {
-        System -> isSystemInDarkMode
-        Light -> false
-        Dark -> true
-    }
-}
-
-enum class WideNotationOption {
-    WideWithW, WideWithLowercase;
-
-    override fun toString() = when (this) {
-        WideWithW -> "Use w (Rw)"
-        WideWithLowercase -> "Use lowercase (r)"
-    }
-}
