+++ app/src/main/java/com/rickyhu/hushkeyboard/settings/options/ThemeOption.kt
@@ -1,0 +1,11 @@
+package com.rickyhu.hushkeyboard.settings.options
+
+enum class ThemeOption {
+    System, Light, Dark;
+
+    fun isDarkTheme(isSystemInDarkMode: Boolean) = when (this) {
+        System -> isSystemInDarkMode
+        Light -> false
+        Dark -> true
+    }
+}
