+++ app/src/main/java/com/rickyhu/hushkeyboard/ui/settings/composables/ThemeOptionDropdownItem.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.rickyhu.hushkeyboard.R
-import com.rickyhu.hushkeyboard.settings.ThemeOption
+import com.rickyhu.hushkeyboard.settings.options.ThemeOption
 import com.rickyhu.hushkeyboard.ui.theme.HushKeyboardTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
